LoadImageOnForm moved to MainImageParams
