+++ KnittingAssistant/ViewModel/Helper/UserControlParams.cs
@@ -1,9 +1,4 @@
 ﻿using KnittingAssistant.View.userControls;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace KnittingAssistant.ViewModel
 {
